+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-### 4.4.2
+### 4.4.2-rc.1
 #### Remove deprecated code
 - Refactor TopicLandingPage
   - Remove warning and only re-render when slug changed
@@ -9,6 +9,9 @@
 - Remove unused article-v1 components
 - Use normalize.css with static assets rather than package
 - Move article container to main chunk
+####  Fix meta data of video reported by Google Search Console #1425 
+- Add thumbnailUrl, description, and uploadDate to metadata of leading video of topic
+- Rename and move leading components to topic
 
 ## Release
 ### 4.4.1